extract test/coverage data loading
