+++ src/createTestReport.ts
@@ -1,39 +1,11 @@
-import { createJUnitParser, type TestSuites } from './utilities/junit_parser.ts';
-import { createLcovSummary, type LcovFile, lcovParser, type LcovSummary } from './utilities/lcov_parser.ts';
 import { convertTestresultsToMarkdown } from './testReportToMarkdown.ts';
 import { convertTestresultsToManifest } from './testReportToManifest.ts';
-import { type TestReportConfig, testReportConfigSchema } from './testReportConfig.ts';
+import type { TestReportConfig } from './testReportConfig.ts';
 import { createBadgeSvg } from './utilities/createBadgeSvg.ts';
-import { readFile } from './utilities/dependencies.ts';
 import { exportOutput } from './utilities/miscUtils.ts';
-
-async function getJUnitData(reportConfig: TestReportConfig): Promise<TestSuites> {
-  // Load all JUnit files and convert and merge them
-  const jUnitParser = createJUnitParser();
-  const result: TestSuites = { name: '', tests: 0, failures: 0, errors: 0, time: 0, testSuites: [] };
-
-  const junitData: TestSuites[] = [];
-  for (const junitFilename of reportConfig.test_results.junit) {
-    try {
-      const junitBytes = await readFile(junitFilename);
-      const xmlData = new TextDecoder().decode(junitBytes);
-      junitData.push(jUnitParser(xmlData));
-    } catch (_error) {
-      // Notify error by adding a test suite with 1 test and 1 error
-      result.tests++;
-      result.errors++;
-    }
-  }
-  // Combine all JUnit data
-  return junitData.reduce((acc, data) => {
-    acc.tests += data.tests;
-    acc.failures += data.failures;
-    acc.errors += data.errors;
-    acc.time += data.time;
-    acc.testSuites.push(...data.testSuites);
-    return acc;
-  }, result);
-}
+import type { LcovSummary } from './utilities/lcov_parser.ts';
+import type { TestSuites } from './utilities/junit_parser.ts';
+import { getTestReportData } from './testReportData.ts';
 
 function createTestBadge(jUnitData: TestSuites, reportConfig: TestReportConfig): string {
   // Collect data
@@ -87,34 +59,8 @@ function createCoverageBadge(lcovSummary: LcovSummary, reportConfig: TestReportC
   });
 }
 
-async function loadReportConfig(reportDefinitionFilename: string): Promise<TestReportConfig> {
-  const reportConfigBytes = await readFile(reportDefinitionFilename);
-  const reportConfigText = new TextDecoder().decode(reportConfigBytes);
-  return testReportConfigSchema.parse(JSON.parse(reportConfigText));
-}
-
-async function loadLcovData(lcovFilenames: string[]): Promise<LcovFile[]> {
-  const lcovDatas: LcovFile[] = [];
-  for (const lcovFilename of lcovFilenames) {
-    try {
-      const lcovBytes = await readFile(lcovFilename);
-      const lcovData = new TextDecoder().decode(lcovBytes);
-      lcovDatas.push(...lcovParser(lcovData));
-    } catch (error) {
-      console.error(`Error reading LCOV file ${lcovFilename}: ${error}`);
-    }
-  }
-  return lcovDatas;
-}
-
 export async function createTestReport(reportDefinitionFilename: string) {
-  // Get the config file and validate it's content
-  const reportConfig = await loadReportConfig(reportDefinitionFilename);
-
-  // Load all LCOV and JUnit files
-  const lcovDatas = await loadLcovData(reportConfig.test_results.coverage);
-  const lcovSummary = createLcovSummary(lcovDatas);
-  const jUnitData = await getJUnitData(reportConfig);
+  const { reportConfig, jUnitData, lcovDatas, lcovSummary } = await getTestReportData(reportDefinitionFilename);
 
   // Create the outputs
   await exportOutput(
