+++ src/utilities/miscUtils.ts
@@ -2,6 +2,7 @@ import { writeFile } from './dependencies.ts';
 
 export const percentage = (value: number, total: number) => value ? `${((value / total) * 100).toFixed(1)}%` : '';
 
+// todo: use @std:path
 export const extractFilename = (filename: string) => filename.replace(/\\/g, '/').split('/').at(-1)!;
 
 export async function exportOutput(filename: string | undefined, createData: () => string) {
